+++ mail_deduplicate/deduplicate.py
@@ -21,8 +21,9 @@ import re
 from collections import Counter
 from difflib import unified_diff
 from itertools import combinations
-from operator import attrgetter
+from operator import attrgetter, itemgetter
 from pathlib import Path
+import random
 
 import click
 from boltons.cacheutils import cachedproperty
@@ -182,7 +183,7 @@ class DuplicateSet:
             if hasattr(DuplicateSet, mid):
                 # Apply strategy.
                 logger.debug(f"Call {mid}()...")
-                return set(filter(getattr(self, mid), self.pool))
+                return getattr(self, mid)()
 
         raise NotImplementedError(f"Can't find any of {method_ids!r} methods.")
 
@@ -233,9 +234,9 @@ class DuplicateSet:
 
         logger.info(f"{candidate_count} mail candidates selected for action.")
         self.stats["set_deduplicated"] += 1
-        return [(mail.source_path, mail.mail_id) for mail in selected_mails]
+        return map(attrgetter("source_path", "mail_id"), selected_mails)
 
-    def discard_older(self, mail):
+    def discard_older(self):
         """Discard all older duplicates.
 
         Only keeps the subset sharing the most recent timestamp.
@@ -244,9 +245,9 @@ class DuplicateSet:
             f"Select all mails strictly older than the {self.newest_timestamp} "
             "timestamp..."
         )
-        return mail.timestamp < self.newest_timestamp
+        return {mail for mail in self.pool if mail.timestamp < self.newest_timestamp}
 
-    def discard_oldest(self, mail):
+    def discard_oldest(self):
         """Discard all the oldest duplicates.
 
         Keeps all mail of the duplicate set but those sharing the oldest
@@ -256,9 +257,9 @@ class DuplicateSet:
             f"Select all mails sharing the oldest {self.oldest_timestamp} "
             "timestamp..."
         )
-        return mail.timestamp == self.oldest_timestamp
+        return {mail for mail in self.pool if mail.timestamp == self.oldest_timestamp}
 
-    def discard_newer(self, mail):
+    def discard_newer(self):
         """Discard all newer duplicates.
 
         Only keeps the subset sharing the most ancient timestamp.
@@ -267,9 +268,9 @@ class DuplicateSet:
             f"Select all mails strictly newer than the {self.oldest_timestamp} "
             "timestamp..."
         )
-        return mail.timestamp > self.oldest_timestamp
+        return {mail for mail in self.pool if mail.timestamp > self.oldest_timestamp}
 
-    def discard_newest(self, mail):
+    def discard_newest(self):
         """Discard all the newest duplicates.
 
         Keeps all mail of the duplicate set but those sharing the newest
@@ -279,9 +280,9 @@ class DuplicateSet:
             f"Select all mails sharing the newest {self.newest_timestamp} "
             "timestamp..."
         )
-        return mail.timestamp == self.newest_timestamp
+        return {mail for mail in self.pool if mail.timestamp == self.newest_timestamp}
 
-    def discard_smaller(self, mail):
+    def discard_smaller(self):
         """Discard all smaller duplicates.
 
         Only keeps the subset sharing the biggest size.
@@ -289,9 +290,9 @@ class DuplicateSet:
         logger.info(
             f"Select all mails strictly smaller than {self.biggest_size} bytes..."
         )
-        return mail.size < self.biggest_size
+        return {mail for mail in self.pool if mail.size < self.biggest_size}
 
-    def discard_smallest(self, mail):
+    def discard_smallest(self):
         """Discard all the smallest duplicates.
 
         Keeps all mail of the duplicate set but those sharing the smallest
@@ -301,9 +302,9 @@ class DuplicateSet:
             f"Select all mails sharing the smallest size of {self.smallest_size} "
             "bytes..."
         )
-        return mail.size == self.smallest_size
+        return {mail for mail in self.pool if mail.size == self.smallest_size}
 
-    def discard_bigger(self, mail):
+    def discard_bigger(self):
         """Discard all bigger duplicates.
 
         Only keeps the subset sharing the smallest size.
@@ -311,9 +312,9 @@ class DuplicateSet:
         logger.info(
             f"Select all mails strictly bigger than {self.smallest_size} bytes..."
         )
-        return mail.size > self.smallest_size
+        return {mail for mail in self.pool if mail.size > self.smallest_size}
 
-    def discard_biggest(self, mail):
+    def discard_biggest(self):
         """Discard all the biggest duplicates.
 
         Keeps all mail of the duplicate set but those sharing the biggest
@@ -323,23 +324,31 @@ class DuplicateSet:
             f"Select all mails sharing the biggest size of {self.biggest_size} "
             "bytes..."
         )
-        return mail.size == self.biggest_size
+        return {mail for mail in self.pool if mail.size == self.biggest_size}
 
-    def discard_matching_path(self, mail):
+    def discard_matching_path(self):
         """ Discard all duplicates whose file path match the regexp. """
         logger.info(
             "Select all mails with file path matching the "
             f"{self.conf.regexp.pattern} regexp..."
         )
-        return re.search(self.conf.regexp, mail.path)
+        return {mail for mail in self.pool if re.search(self.conf.regexp, mail.path)}
 
-    def discard_non_matching_path(self, mail):
+    def discard_non_matching_path(self):
         """ Discard all duplicates whose file path doesn't match the regexp. """
         logger.info(
             "Select all mails with file path not matching the "
             f"{self.conf.regexp.pattern} regexp..."
         )
-        return not re.search(self.conf.regexp, mail.path)
+        return {
+            mail for mail in self.pool if not re.search(self.conf.regexp, mail.path)
+        }
+
+    def discard_one(self):
+        return {random.choice(tuple(self.pool))}
+
+    def discard_all_but_one(self):
+        return set(random.sample(self.pool, k=len(self.pool) - 1))
 
 
 class Deduplicate: